Fix compilation on GCC11.3+ which don't typedef uint

+++ src/usb_definitions.h
@@ -86,8 +86,8 @@ typedef struct struct_usb_device_t usb_device_t;
 typedef struct struct_root_port_t {
   volatile bool initialized;
   volatile bool addr0_exists;
-  volatile uint pin_dp;
-  volatile uint pin_dm;
+  volatile uint8_t pin_dp;
+  volatile uint8_t pin_dm;
   volatile usb_device_event_t event;
   usb_device_t *root_device;
